linter fixes
