+++ Sources/SwiftletModel/Model/EntityModel.swift
@@ -8,6 +8,7 @@
 import Foundation
 
 public protocol EntityModel {
+    // swiftlint:disable:next type_name
     associatedtype ID: Hashable & Codable & LosslessStringConvertible
 
     var id: ID { get }
